+++ lib/rex/post/channel.rb
@@ -1,0 +1,4 @@
+# -*- coding: binary -*-
+
+require 'rex/post/channel/container'
+require 'rex/post/channel/socket_abstraction'
